Form Validation
Added some form validation to the create a contact page.
Validation is live as the user types or when the form is submitted.
Form will not submit if the data is in the wrong format or required data
is missing.
A red border and error message will appear with the invalid fields.